requirements.txt
streamlit
gradio~=5.34.2
transformers~=4.52.4
accelerate
torch~=2.7.1
diffusers
Pillow~=11.2.1
openai~=1.91.0
scikit-learn
matplotlib
plotly
python-dotenv
easyocr~=1.7.2
numpy~=2.3.1
pytesseract~=0.3.13
langchain~=0.3.26
boto3~=1.38.44
pdf2image~=1.17.0